fix: Refactor and fix the broken imports
