+++ dataspace/managers/auth_manager.py
