+++ dataspace/setup.py
@@ -25,7 +25,7 @@
 from setuptools import setup, find_packages
 
 setup(
-    name="tx_dataspace_sdk",
+    name="dspace",
     version="0.1.0",
     packages=find_packages(),
 )
