+++ dataspace/tests/manager/__init__.py
@@ -24,5 +24,3 @@
 # Package-level variables
 __author__ = 'Eclipse Tractus-X Contributors'
 __license__ = "Apache License, Version 2.0"
-
-from .auth_manager import AuthManager
