+++ dataspace/tests/manager/test_auth_manager.py
@@ -24,7 +24,7 @@ import unittest
 from unittest.mock import Mock
 from fastapi import Request
 from starlette.datastructures import Headers
-from tractusx_sdk.dataspace.managers.auth_manager import AuthManager
+from managers.auth_manager import AuthManager
 
 class TestAuthManager(unittest.TestCase):
