+++ industry/__init__.py
@@ -24,5 +24,3 @@
 __version__ = '0.0.1'
 __author__ = 'Eclipse Tractus-X Contributors'
 __license__ = "Apache License, Version 2.0"
-
-
