+++ industry/services/aas_service.py
